requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
comet-ml==3.47.3
configobj==5.0.9
dulwich==0.22.4
everett==3.1.0
filelock==3.16.1
ftfy==6.2.3
gdown==5.2.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
libtmux==0.37.0
markdown-it-py==3.0.0
mdurl==0.1.2
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.2
pillow==10.4.0
pkgutil_resolve_name==1.3.10
psutil==6.1.0
Pygments==2.18.0
PySocks==1.7.1
python-box==6.1.0
PyWavelets==1.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==2.18.0
simplejson==3.19.3
soupsieve==2.6
threadpoolctl==3.5.0
tifffile==2023.7.10
tmuxp==1.47.0
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.16.0
wurlitzer==3.1.1
zipp==3.20.2
